States can now change to a new state at the end of a run

+++ common/ledimu/src/imurunner.cpp
@@ -8,7 +8,7 @@ void ImuRunner::set_all(const Color& color)
 {
 	for(int i=0;i<m_pixel_count;i++) 
 	{
-		std::cout << "Setting pixel " << i << " to " << color.red << ", " << color.green << ", " << color.blue << std::endl;
+		std::cout << "Setting pixel " << i << " to " << (int)color.red << ", " << (int)color.green << ", " << (int)color.blue << std::endl;
 		set_pixel(i, color);
 	}
 }
